# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Role:
    enum:
    - user
    - admin
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleAdmin
...
  domain.UpdateUserRequest:
    properties:
      alamat:
        example: Jl. Contoh No. 123
        type: string
      nama:
        example: John Doe
        type: string
      no_hp:
        example: "08123456789"
        type: string
      password:
        example: newpassword123
        type: string
    type: object
...
  domain.UserResponse:
    properties:
      alamat:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      nama:
        type: string
      no_hp:
        type: string
      role:
        $ref: '#/definitions/domain.Role'
      updated_at:
        type: string
    type: object
...
  utils.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  utils.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/utils.Meta'
      success:
        type: boolean
    type: object
  utils.StandardResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@jubel.app
    name: Jubel API Support
  description: API untuk aplikasi jual beli barang bekas Jubel
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Jubel API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Mendapatkan daftar pengguna dengan paginasi (admin only)
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.UserResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus pengguna berdasarkan ID (self atau admin)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Mendapatkan detail pengguna berdasarkan ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Memperbarui data pengguna
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /users/me:
    get:
      consumes:
      - application/json
      description: Mendapatkan detail pengguna yang sedang login
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"